Moved geometry encoding to a dedicated module
