+++ mapbox_vector_tile/encoder.py
@@ -14,18 +14,7 @@ from shapely.wkb import loads as load_wkb
 from shapely.wkt import loads as load_wkt
 import decimal
 from .compat import PY3, vector_tile, apply_map
-
-
-# tiles are padded by this number of pixels for the current zoom level
-padding = 0
-
-cmd_bits = 3
-tolerance = 0
-
-CMD_MOVE_TO = 1
-CMD_LINE_TO = 2
-CMD_SEG_END = 7
-CMD_FAKE = 0
+from .geom_encoder import GeometryEncoder
 
 
 def on_invalid_geometry_raise(shape):
@@ -306,185 +295,6 @@ class VectorTile:
 
                 feature.tags.append(self.seen_values_idx[v])
 
-    def _handle_skipped_last(self, f, skipped_index, cur_x, cur_y, x_, y_):
-        last_x = f.geometry[skipped_index - 2]
-        last_y = f.geometry[skipped_index - 1]
-        last_dx = ((last_x >> 1) ^ (-(last_x & 1)))
-        last_dy = ((last_y >> 1) ^ (-(last_y & 1)))
-        dx = cur_x - x_ + last_dx
-        dy = cur_y - y_ + last_dy
-        x_ = cur_x
-        y_ = cur_y
-        f.geometry.__setitem__(skipped_index - 2, ((dx << 1) ^ (dx >> 31)))
-        f.geometry.__setitem__(skipped_index - 1, ((dy << 1) ^ (dy >> 31)))
-
-    def _parseGeometry(self, shape):
-        coordinates = []
-        lineType = "line"
-        polygonType = "polygon"
-
-        def _get_arc_obj(arc, type):
-            # seq = ( (x, y, CMD_LINE_TO) for (x, y) in islice(arc.coords, 1, len(arc.coords)))
-            # coordinates.extend(seq)
-
-            # x, y = arc.coords[0]
-            # coordinates.append((x, y, CMD_MOVE_TO))
-            # seq = ( (x, y, CMD_LINE_TO) for (x, y) in islice(arc.coords, 1, len(arc.coords)))
-            # coordinates.extend(seq)
-            # if type == polygonType:
-            #     cmd = CMD_MOVE_TO
-            # else:
-            #     cmd = CMD_SEG_END
-            # x, y = arc.coords[-1]
-            # coordinates.append((x, y, cmd))
-            #
-            #
-            cmd = CMD_MOVE_TO
-            length = len(arc.coords)
-            for i, (x, y) in enumerate(arc.coords):
-                if i == 0:
-                    cmd = CMD_MOVE_TO
-                elif i == length - 1 and type == polygonType:
-                    cmd = CMD_SEG_END
-                else:
-                    cmd = CMD_LINE_TO
-                coordinates.append((x, y, cmd))
-
-        if shape.type == 'GeometryCollection':
-            # do nothing
-            coordinates = []
-
-        elif shape.type == 'Point':
-            coordinates.append((shape.x, shape.y, CMD_MOVE_TO))
-
-        elif shape.type == 'LineString':
-            _get_arc_obj(shape, lineType)
-
-        elif shape.type == 'Polygon':
-            rings = [shape.exterior] + list(shape.interiors)
-            for ring in rings:
-                _get_arc_obj(ring, polygonType)
-
-        elif shape.type == 'MultiPoint':
-            coordinates += [(point.x, point.y, CMD_MOVE_TO)
-                            for point in shape.geoms]
-
-        elif shape.type == 'MultiLineString':
-            for arc in shape.geoms:
-                _get_arc_obj(arc, lineType)
-
-        elif shape.type == 'MultiPolygon':
-            for polygon in shape.geoms:
-                rings = [polygon.exterior] + list(polygon.interiors)
-                for ring in rings:
-                    _get_arc_obj(ring, polygonType)
-
-        else:
-            raise NotImplementedError("Can't do %s geometries" % shape.type)
-
-        return coordinates
-
-    def _geo_encode_old(self, f, shape, y_coord_down):
-        x_, y_ = 0, 0
-
-        cmd = -1
-        cmd_idx = -1
-        vtx_cmd = -1
-        prev_cmd = -1
-
-        skipped_index = -1
-        skipped_last = False
-        cur_x = 0
-        cur_y = 0
-
-        it = 0
-        length = 0
-
-        coordinates = self._parseGeometry(shape)
-
-        while it < len(coordinates):
-            x, y, vtx_cmd = coordinates[it]
-
-            if vtx_cmd != cmd:
-                if cmd_idx >= 0:
-                    f.geometry[cmd_idx] = self._encode_cmd_length(cmd, length)
-
-                cmd = vtx_cmd
-                length = 0
-                cmd_idx = len(f.geometry)
-                f.geometry.append(0)  # placeholder added in first pass
-
-            if (vtx_cmd == CMD_MOVE_TO or vtx_cmd == CMD_LINE_TO):
-                if cmd == CMD_MOVE_TO and skipped_last and skipped_index > 1:
-                    self._handle_skipped_last(
-                        f, skipped_index, cur_x, cur_y, x_, y_)
-
-                # ensure that floating point values don't get truncated
-                if isinstance(x, float):
-                    x = self._round(x)
-                if isinstance(y, float):
-                    y = self._round(y)
-
-                x = int(x)
-                y = int(y)
-
-                if not y_coord_down:
-                    y = self.extents - y
-
-                # Compute delta to the previous coordinate.
-                cur_x = int(x)
-                cur_y = int(y)
-
-                dx = cur_x - x_
-                dy = cur_y - y_
-
-                sharp_turn_ahead = False
-
-                if (it + 2 <= len(coordinates)):
-                    next_x, next_y, next_cmd = coordinates[it + 1]
-                    if next_cmd == CMD_LINE_TO:
-                        next_dx = fabs(cur_x - int(next_x))
-                        next_dy = fabs(cur_y - int(next_y))
-                        if ((next_dx == 0 and next_dy >= tolerance) or
-                                (next_dy == 0 and next_dx >= tolerance)):
-                            sharp_turn_ahead = True
-
-                # Keep all move_to commands, but omit other movements
-                # that are not >= the tolerance threshold and should
-                # be considered no-ops.
-                # NOTE: length == 0 indicates the command has changed and will
-                # preserve any non duplicate move_to or line_to
-                if (length == 0 or sharp_turn_ahead or
-                        fabs(dx) >= tolerance or fabs(dy) >= tolerance):
-                    # Manual zigzag encoding.
-                    f.geometry.append((dx << 1) ^ (dx >> 31))
-                    f.geometry.append((dy << 1) ^ (dy >> 31))
-                    x_ = cur_x
-                    y_ = cur_y
-                    skipped_last = False
-                    length = length + 1
-                else:
-                    skipped_last = True
-                    skipped_index = len(f.geometry)
-            elif vtx_cmd == CMD_SEG_END:
-                if prev_cmd != CMD_SEG_END:
-                    length = length + 1
-            else:
-                raise Exception("Unknown command type: '%s'" % vtx_cmd)
-
-            it = it + 1
-            prev_cmd = cmd
-
-        # at least one vertex + cmd/length
-        if skipped_last and skipped_index > 1:
-            # if we skipped previous vertex we just update it to the
-            # last one here.
-            self._handle_skipped_last(f, skipped_index, cur_x, cur_y, x_, y_)
-
-        # Update the last length/command value.
-        if cmd_idx >= 0:
-            f.geometry[cmd_idx] = self._encode_cmd_length(cmd, length)
-
     def _geo_encode(self, f, shape, y_coord_down):
 
         geom_encoder = GeometryEncoder(f.geometry, y_coord_down, self.extents, self._round)
@@ -505,126 +315,5 @@ class VectorTile:
             geom_encoder.encode_polygon(shape)
         elif shape.type == 'MultiPolygon':
             geom_encoder.encode_polygons(shape)
-        # print str(f.geometry)
-
-
-def omit_last(iterator):
-    try:
-        next_elmt = next(iterator)
-        while True:
-            elmt = next_elmt
-            next_elmt = next(iterator)
-            yield elmt
-    except StopIteration:
-        pass
-
-
-class GeometryEncoder:
-
-    def __init__(self, geometry, y_coord_down, extents, round_fn):
-        self._geometry = geometry
-        self._y_coord_down = y_coord_down
-        self._extents = extents
-        self._round = round_fn
-        self._cmd_idx = -1
-        self._geom_size = 0
-        self.last_x = 0
-        self.last_y = 0
-
-    def _encode_cmd_length(self, cmd, length):
-        return (length << cmd_bits) | (cmd & ((1 << cmd_bits) - 1))
-
-    def reserve_space_for_cmd(self):
-        # print("Reserve space for command at pos {}".format(len(f.geometry)))
-        self._geometry.append(CMD_FAKE)
-        self._cmd_idx = self._geom_size
-        self._geom_size += 1
-
-    def append_cmd(self, cmd, length):
-        cmd_encoded = self._encode_cmd_length(cmd, length)
-        # print("Command {} at pos {}".format(cmd_encoded, len(f.geometry)))
-        self._geometry.append(cmd_encoded)
-        self._geom_size += 1
-
-    def set_back_cmd(self, cmd):
-        length = (self._geom_size - self._cmd_idx) >> 1
-        cmd_encoded = self._encode_cmd_length(cmd, length)
-        # print("Command back {} , length {} at pos {}".format(cmd_encoded, length, self._cmd_idx))
-        self._geometry[self._cmd_idx] = cmd_encoded
-
-    def append_coords(self, fx, fy, force=False):
-        # print("Append coords {}, {} at pos {}".format(fx, fy, len(f.geometry)))
-        if isinstance(fx, float):
-            x = int(self._round(fx))
         else:
-            x = fx
-        if isinstance(fy, float):
-            y = int(self._round(fy))
-        else:
-            y = fy
-        if not self._y_coord_down:
-            y = self._extents - y
-
-        dx = x - self.last_x
-        dy = y - self.last_y
-        if not force and dx == 0 and dy == 0:
-            # print("...Aborted")
-            return
-        self._geometry.append((dx << 1) ^ (dx >> 31))
-        self._geometry.append((dy << 1) ^ (dy >> 31))
-        self.last_x = x
-        self.last_y = y
-        self._geom_size += 2
-
-    def append_ring(self, arc):
-        it = iter(arc.coords)
-        x, y = next(it)
-        self.append_cmd(CMD_MOVE_TO, 1)
-        self.append_coords(x, y, True)
-        self.reserve_space_for_cmd()
-        for x, y in omit_last(it):
-            self.append_coords(x, y)
-        self.set_back_cmd(CMD_LINE_TO)
-        self.append_cmd(CMD_SEG_END, 1)
-
-    def append_arc(self, arc):
-        it = iter(arc.coords)
-        x, y = next(it)
-        self.append_cmd(CMD_MOVE_TO, 1)
-        self.append_coords(x, y, True)
-        self.reserve_space_for_cmd()
-        for x, y in it:
-            self.append_coords(x, y)
-        self.set_back_cmd(CMD_LINE_TO)
-
-    def append_polygon(self, shape):
-        self.append_ring(shape.exterior)
-        for arc in shape.interiors:
-            self.append_ring(arc)
-
-    def encode_point(self, shape):
-        self.append_cmd(CMD_MOVE_TO, 1)
-        self.append_coords(shape.x, shape.y, True)
-
-    def encode_points(self, shape):
-        self.append_cmd(CMD_MOVE_TO, len(shape.geoms))
-        # map ? apply ?
-        for point in shape.geoms:
-            self.append_coords(point.x, point.y, True)
-
-    def encode_linestring(self, shape):
-        self.append_arc(shape)
-
-    def encode_linestrings(self, shape):
-        for arc in shape.geoms:
-            self.append_arc(arc)
-
-    def encode_polygon(self, shape):
-        self.append_polygon(shape)
-
-    def encode_polygons(self, shape):
-        for polygon in shape.geoms:
-            self.append_polygon(polygon)
-
-
-
+            raise NotImplementedError("Can't do %s geometries" % shape.type)
